Split wave group recovery helpers
