+++ components/waves/specs/groups/group/edit/buttons/utils/waveGroupCloneRecovery.ts
@@ -1,0 +1,34 @@
+import type { ApiWave } from "@/generated/models/ApiWave";
+import { commonApiFetch } from "@/services/api/common-api";
+
+const waveReferencesGroup = (wave: ApiWave, groupId: string): boolean =>
+  [
+    wave.visibility.scope.group?.id,
+    wave.participation.scope.group?.id,
+    wave.voting.scope.group?.id,
+    wave.chat.scope.group?.id,
+    wave.wave.admin_group.group?.id,
+  ].includes(groupId);
+
+export const getCloneReferenceState = async ({
+  waveId,
+  groupId,
+}: {
+  readonly waveId: string;
+  readonly groupId: string;
+}): Promise<"attached" | "unattached" | "unknown"> => {
+  try {
+    const currentWave = await commonApiFetch<ApiWave>({
+      endpoint: `waves/${waveId}`,
+    });
+    return waveReferencesGroup(currentWave, groupId)
+      ? "attached"
+      : "unattached";
+  } catch (error) {
+    console.warn(
+      "[WaveGroupEditButtons] Unable to verify cloned group attachment",
+      { waveId, groupId, error }
+    );
+    return "unknown";
+  }
+};
